Extract FakeGameView to use in multiple tests
